rag-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langchain-community==0.3.27
chromadb==1.0.16
sentence-transformers==5.1.0
pypdf==4.2.0
python-dotenv==1.1.1